+++ client/src/consts.rs
@@ -4,5 +4,5 @@ pub const LAMPORTS_PER_TX: u64    = 5000;
 pub const DECIMAL_PLACES: u8      = 6;
 pub const PURCHASE_CAP: u64       = 5_000_000_000; // 5,000 USDC with 6 decimals
 pub const SALE_CAP: u64           = 1_000_000_000; // 1,000 tokens with 6 decimals
-pub const BUY_FEE_BPS: u32        = 0; // 0% fee
-pub const SELL_FEE_BPS: u32       = 100; // 1% fee
+pub const BUY_FEE_BPS: u16        = 0; // 0% fee
+pub const SELL_FEE_BPS: u16       = 100; // 1% fee
